test client: reorder connection_handle_data()
Process HUP as the last thing, as it may be possible to get both HUP and
IN at the same time: the remote end close the connection but the buffers
still have unread data. This makes sure we read all the data after a
HUP.

It is also assumes that you cannot have HUP and OUT at the same time,
since probably trying to write after HUP would cause an error.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.co.uk>