test client: provoke protocol errors
The server does not yet implement wthp_region, so trying to create one
will lead to a protocol error. This will get fixed once wtph_region is
implemented.

Since we don't have error handling yet, plug in some ad hoc reporting so
we at least see the messages. This serves as a reminder to improve
things.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.co.uk>