Merge libwalthamclient and libwalthamserver
There is no need to separate them, as almost all code is common
and there is no reason an application couldn't be both a server
and a client at the same time. This also solves some conflicting
symbols.

Eventually we plan to stop compiling the marshalling and demarshalling
functions into the shared library, and then libwaltham will be even
more streamlined.

Signed-off-by: Emilio Pozuelo Monfort <emilio.pozuelo@collabora.co.uk>