Move in wth_connection_roundtrip() implementation
The test client already had a pretty generic implementation of a
roundtrip. Move it into Waltham proper as a helper.

The roundtrip could be implemented perfectly fine outside of Waltham
(where it was), but offering a helper for it seems like a nice thing to
do. The documentation recommends to do a roundtrip as the last thing in
a normal client shutdown, so a ready-made function is nice to have.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.co.uk>