Add LICENSE
This is the license used by all source code.

The license text is taken from waltham-connection.h. The text is exactly
the same in all .c, .h, and .txt files (that have a license blurb) to
the letter. The following files contain the same text with just
differently split lines:
	w-util.c
	w-util.h
	waltham-private.h
	waltham-util.c
	waltham-util.h
	command.xml
	ivi-application.xml

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.co.uk>
Reviewed-by: Emilio Pozuelo Monfort <emilio.pozuelo@collabora.co.uk>