waltham: add logging functions
These are the logging API with three levels: debug, error, and abort
which also kills the program. The actual implementation can be refined
later to e.g. include timestamp and allow debug output be optional.

The style here was chosen so that the format string should not include a
newline as one will be added automatically.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.co.uk>