connection: replace perror() with wth_debug()
For consistency on logging, if we e.g. want timestamps, redirection,
etc.

The strerror() usage is not thread-safe, but fixing that is for another
day as the situation with portable and thread-safe strerror()
replacement is dire.

Add checking of the dispatch call in wth_connection_roundtrip(). Now the
debug output will tell if there was a protocol error.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.co.uk>