waltham: assert server/client side
This adds the helpers for asserting that a function is called on the
side it is intended on. Otherwise the program gets aborted, because it's
a programmer error.

wth_connection_roundtrip() is marked as client-only.

wth_object_post_error() is marked as server-only.

The API produced by the generator should get a similar treatment.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.co.uk>