Move the wth_display implementations into waltham-connection
This effectively makes the global wth_display object private to the
wth_connection, which is now exposing the parts relevant for clients and
servers through exported functions.

This allows internal tracking of object ids, display versioning being
negotiated transparently, and generally a much better control over the
internals of the protocol.

Signed-off-by: Emmanuel Gil Peyrot <emmanuel.peyrot@collabora.com>